requirements.txt
# torch==2.4.0
# torchvision==0.19.0
accelerate==0.32.0
diffusers==0.24.0
matplotlib
# numpy==1.26.4
opencv_python==4.10.0.84
pillow==10.3.0
PyYAML==6.0.1
scipy==1.13.1
scikit-image==0.24.0
tqdm==4.66.4
transformers==4.46.3
fvcore==0.1.5.post20221221
cloudpickle==3.0.0
omegaconf==2.3.0
pycocotools==2.0.8
av==12.3.0
gradio==4.43.0
peft==0.17.0
huggingface_hub==0.25.0
python-dotenv>=1.0.0
requests>=2.31.0
carvekit
config==0.5.1
einops
onnxruntime
# pydantic==2.8.2
mediapipe==0.10.21
scikit-learn
ultralytics
git+https://github.com/xinntao/BasicSR.git
git+https://github.com/ultralytics/CLIP.git